caso 7

+++ src/projeto/hospital/gerencia/prontuario/paciente/fidelidade/Fidelidade.java
@@ -6,14 +6,14 @@ package projeto.hospital.gerencia.prontuario.paciente.fidelidade;
  * @author Estacio Pereira
  *
  */
-public interface Fidelidade {
+public interface Fidelidade{
 
 	/**
 	 * Retorna o desconto dos pacientes nos servicos.
 	 * 
 	 * @return Porcentagem do desconto em servicos do SOOS.
 	 */
-	public double getDescontoServico();
+	public Double getDescontoServico();
 
 	/**
 	 * Retorna a porcentagem de bonus que um paciente recebe em relacao aos
@@ -21,5 +21,5 @@ public interface Fidelidade {
 	 * 
 	 * @return Porcentagem do bonus de pontos no SOOS.
 	 */
-	public double getCreditoBonus();
+	public Integer getCreditoBonus();
 }
